FullRecipe kind of done
Also added description to dishView.

But it would be fun to add a real description and instruction and also
more dishes to Model.

Perhaps the ingredients should show in the Full recipe?

Also, there's a bug when selecting main course, 2 main courses can be
added to the dinner.

And the button for printing full recipe is misplaced.