Fixed things here and there
Should be ok to hand in now.

Changed so the correct views appear.
Still a bit buggy that the DishList is at first not visible and that the
preloaded dishes are not visible in the cart until participants are
changed...

Added some dishes. Changed some styling.
Maid the first view window draggable, check it out. =)

There is still a bug when adding a "main dish" with model.addDishToMenu.
Since there is already a main dish, somehow, it's still possible to add
another! Could it be the space....?